the cart and the category  are ready

+++ demo/demo/src/main/java/com/example/demo/persistence/service/CartService.java
@@ -2,6 +2,7 @@ package com.example.demo.persistence.service;
 
 
 import com.example.demo.persistence.entity.Cart;
+import com.example.demo.persistence.entity.User;
 import com.example.demo.persistence.repository.CartRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,15 +14,29 @@ public class CartService {
     @Autowired
     private CartRepository cartRepository;
 
-    public Optional<Cart> findById(Long id) {
-            return cartRepository.findById(id);
-        }
-
     public Cart save(Cart cart) {
         return cartRepository.save(cart);
     }
+    public Optional<Cart> findById(Long id) {
+        Optional<Cart> optionalCart = cartRepository.findById(id);
+        if (!optionalCart.isPresent()) {
+            throw new CartService.CartNotFoundException("Cart not found with id: " + id);
+        }
+        return cartRepository.findById(id);
+        }
 
     public void deleteById(Long id) {
+
+        Optional<Cart> optionalCart = cartRepository.findById(id);
+        if (!optionalCart.isPresent()) {
+            throw new CartService.CartNotFoundException("Cart not found with id: " + id);
+        }
         cartRepository.deleteById(id);
+    }
+
+    public class CartNotFoundException extends RuntimeException{
+        public CartNotFoundException(String message){
+            super(message);
+        }
     }
 }
